refactor: split out svg icon creation
